+++ migrations/900_assessment_access_rules__exam_uuid__add.sql
@@ -1,0 +1,2 @@
+ALTER TABLE exams ADD COLUMN IF NOT EXISTS uuid uuid NULL UNIQUE;
+ALTER TABLE assessment_access_rules ADD COLUMN exam_uuid uuid REFERENCES exams(uuid) ON UPDATE CASCADE ON DELETE SET NULL;