+++ ChangeLog.md
@@ -25,6 +25,8 @@
 
   * Change external grading to receive results from an SQS queue instead of a webhook (Nathan Walters).
 
+  * Change assessment access rule examId linking to examUuid (Dave Mussulman).
+
   * Fix load-reporting close during unit tests (Matt West).
 
   * Fix PL / scheduler linking stored procedure to allow linked exams and fix bugs (Dave Mussulman).